requirements.txt
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
python-dateutil==2.8.2
pytest==7.4.0
pyyaml==6.0.1
